refactor(db_handler): Restructured imports
